// src/app/models/ReportModel.ts
export class BPCReportFGCPS extends CommonClass {
    ID: number;
    Client: string;
    Company: string;
    Type: string;
    PatnerID: string;
    Plant: string;
    Material: string;
    MaterialText: string;
    StickQty: any;
    UOM: string;
    Batch: string;
    Price: any;
    Status: string;
}